Add wallet-rescan.md, add helper method to WalletApi that allows you to clear out the wallet's utxos/addresses
