+++ testkit/src/main/resources/reference.conf
@@ -3,7 +3,7 @@ bitcoin-s {
     network = regtest # regtest, testnet3, mainnet
 
     logging {
-        level = WARN # trace, debug, info, warn, error, off
+        level = INFO # trace, debug, info, warn, error, off
 
         # You can also tune specific module loggers.
         # They each take the same levels as above.
