+++ website/sidebars.json
@@ -29,7 +29,8 @@
       "wallet/wallet",
       "wallet/chain-query-api",
       "wallet/node-api",
-      "wallet/dlc"
+      "wallet/dlc",
+      "wallet/wallet-rescan"
     ],
     "RPC Clients": [
       "rpc/rpc-clients-intro",
